requirements/_base.txt
#https://www.djangoproject.com/download/1.5c1/tarball/
django==1.4.3
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
unicodecsv==0.9.0
wsgiref==0.1.2
